pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.0"
description = "Difference generator"
authors = ["Vasilii <petgga@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
flake8 = "^5.0.4"
pip = "^22.2.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"